requirements.txt
# AEGIS Dependencies
# LangGraph Stack (v2.1)

# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# LangGraph + AWS Bedrock
langgraph>=0.0.30
langchain>=0.1.0
langchain-aws>=0.1.0
langchain-community>=0.0.20

# PII Scrubbing (Security)
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
spacy>=3.7.0

# Vector Store
chromadb>=0.4.22

# Redis
redis>=5.0.0
hiredis>=2.3.0

# HTTP Client
httpx>=0.26.0

# AWS (for SSM remediation)
boto3>=1.34.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6
PyYAML>=6.0
schedule>=1.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
